chore(전력, 가격) : 모든 전력, 가격 필드의 자료형 double로 수정
스프링 부트 실행 후 websocket 메시지 및 api 동작 확인까지 완료

+++ src/main/java/com/example/quaterback/api/domain/txinfo/domain/TransactionInfoDomain.java
@@ -19,8 +19,8 @@ public class TransactionInfoDomain {
     private String idToken;
     private String stationId;
     private Integer evseId;
-    private Integer totalMeterValue;
-    private Integer totalPrice;
+    private Double totalMeterValue;
+    private Double totalPrice;
 
     public static TransactionInfoDomain fromStartedTxEventDomain(TransactionEventDomain domain, String stationId) {
         return TransactionInfoDomain.builder()
@@ -33,7 +33,7 @@ public class TransactionInfoDomain {
                 .build();
     }
 
-    public static TransactionInfoDomain fromEndedTxEventDomain(TransactionEventDomain domain, Integer totalMeterValue, Integer totalPrice) {
+    public static TransactionInfoDomain fromEndedTxEventDomain(TransactionEventDomain domain, Double totalMeterValue, Double totalPrice) {
         return TransactionInfoDomain.builder()
                 .transactionId(domain.extractTransactionId())
                 .endedTime(domain.getTimestamp())
